8270646: Improved scanning of XML entities
